+++ spec/service_actor/result_spec.rb
@@ -1,8 +1,9 @@
 # frozen_string_literal: true
 
 RSpec.describe ServiceActor::Result do
+  let(:result) { described_class.new }
+
   it "defines a method ending with *? suffix for each attribute" do
-    result = described_class.new
     result.name = "Sunny"
 
     expect(result.respond_to?(:name?)).to be true
@@ -11,7 +12,6 @@ RSpec.describe ServiceActor::Result do
   context "when input is String" do
     context "when is empty" do
       it "returns false" do
-        result = described_class.new
         result.name = ""
 
         expect(result.name?).to be false
@@ -20,7 +20,6 @@ RSpec.describe ServiceActor::Result do
 
     context "when is not empty" do
       it "returns true" do
-        result = described_class.new
         result.name = "Actor"
 
         expect(result.name?).to be true
@@ -31,7 +30,6 @@ RSpec.describe ServiceActor::Result do
   context "when input is Array" do
     context "when is empty" do
       it "returns false" do
-        result = described_class.new
         result.options = []
 
         expect(result.options?).to be false
@@ -40,7 +38,6 @@ RSpec.describe ServiceActor::Result do
 
     context "when is not empty" do
       it "returns true" do
-        result = described_class.new
         result.options = [1, 2, 3]
 
         expect(result.options?).to be true
@@ -51,7 +48,6 @@ RSpec.describe ServiceActor::Result do
   context "when input is Hash" do
     context "when empty" do
       it "returns false" do
-        result = described_class.new
         result.options = {}
 
         expect(result.options?).to be false
@@ -60,7 +56,6 @@ RSpec.describe ServiceActor::Result do
 
     context "when not empty" do
       it "returns true" do
-        result = described_class.new
         result.options = { name: "Actor" }
 
         expect(result.options?).to be true
@@ -70,7 +65,6 @@ RSpec.describe ServiceActor::Result do
 
   context "when input is NilClass" do
     it "returns false" do
-      result = described_class.new
       result.name = nil
 
       expect(result.name?).to be false
@@ -79,7 +73,6 @@ RSpec.describe ServiceActor::Result do
 
   context "when input is TrueClass" do
     it "returns true" do
-      result = described_class.new
       result.name = true
 
       expect(result.name?).to be true
@@ -88,10 +81,27 @@ RSpec.describe ServiceActor::Result do
 
   context "when input is FalseClass" do
     it "returns true" do
-      result = described_class.new
       result.name = false
 
       expect(result.name?).to be false
+    end
+  end
+
+  describe "#fail!" do
+    it "merges the hash into the result and marks the result as failed" do
+      expect { result.fail!(name: "Sunny") }
+        .to raise_error(ServiceActor::Failure)
+
+      expect(result.name).to eq("Sunny")
+      expect(result).to be_a_failure
+    end
+
+    context "with no arguments" do
+      it "raises and marks the result as failed" do
+        expect { result.fail! }.to raise_error(ServiceActor::Failure)
+
+        expect(result).to be_a_failure
+      end
     end
   end
 end